+++ lib/nito/cp.rb
@@ -2,7 +2,7 @@ require 'fileutils'
 
 module Nito
   class Cp
-    @@pass = nil
+    #@@pass = nil
 
     def initialize(src, dst, perm = 0644)
       @src = src
@@ -11,7 +11,7 @@ module Nito
       if ID == "0"
         root
       else
-        @@pass = Pass.new if ! @@pass
+        #@@pass = Pass.new if ! @@pass
         sudo
       end
     end
@@ -20,14 +20,16 @@ module Nito
 
     def root
       FileUtils.copy_file(@src, @dst)
-      FileUtils.chmod(@perm.to_s, @dst)
+      FileUtils.chmod(@perm, @dst)
     end
 
     def sudo
-      Sudo.run("cp #{@src} #{@dst}", @@pass.secret)
+      #Sudo.run("cp #{@src} #{@dst}", @@pass.secret)
+      Sudo.run("cp #{@src} #{@dst}")
       perm = sprintf "%o", @perm
       #puts "Applying perm #{perm}"
-      Sudo.run("chmod #{perm} #{@dst}", @@pass.secret)
+      #Sudo.run("chmod #{perm} #{@dst}", @@pass.secret)
+      Sudo.run("chmod #{perm} #{@dst}")
     end
   end
 end