+++ lib/nito/sudo.rb
@@ -1,7 +1,8 @@
 module Nito
   module Sudo
     def self.run(command, input = nil)
-      IO.popen("sudo -S #{command}", 'r+') do |io|
+      #IO.popen("sudo -S #{command}", 'r+') do |io|
+      IO.popen("sudo #{command}", 'r+') do |io|
         begin
           io.puts input
           io.close_write